+++ CHANGELOG.md
@@ -18,3 +18,4 @@
 - [fix: memory not released while using wildcard api call with circuitbreaker enabled](https://github.com/Tencent/spring-cloud-tencent/pull/1335)
 - [feat: support 2.0.0](https://github.com/Tencent/spring-cloud-tencent/pull/1458)
 - [feat: support 2.0.0 config](https://github.com/Tencent/spring-cloud-tencent/pull/1463)
+- [feat:upgrade trace plugin.](https://github.com/Tencent/spring-cloud-tencent/pull/1467)